# server/api/v1/static/swagger.json
{
    "openapi": "3.0.0",
  "info": {
    "title": "AEDC Subscription Tracker Application - OpenAPI 3.0",
    "description": "This application is designed to track the subscriptions of the company using the information provided by the users of the application.",
    "contact": {
      "email": "Justin.Ebedi@abujaelectricity.com"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/api/v1/"
    }
  ],
  "tags": [
    {
      "name": "index",
      "description": "All index endpoints"
    },
    {
      "name": "users",
      "description": "All users based endpoints"
    },
    {
      "name": "subscriptions",
      "description": "All subscription based endpoints"
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "tags": [
          "index"
        ],
        "summary": "Check the status of the API",
        "operationId": "Check API",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "status": "OK"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Retrieves the list of all users",
        "operationId": "get_users",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Missing Authorization Header"
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Gets a user by ID",
        "operationId": "get_user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID used to get user",
            "required": true,
            "schema": {
              "type": "string",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "first_name": {
                      "type": "string"
                    },
                    "last_name": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "email": "james.bond@gmail.com",
                    "first_name": "james",
                    "last_name": "bond",
                    "id": "007"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Deletes a user by ID",
        "operationId": "delete_user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID used to get user",
            "required": true,
            "schema": {
              "type": "string",
              "example": 10
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Updates a user's information",
        "operationId": "update_user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID used to get user",
            "required": true,
            "schema": {
              "type": "string",
              "example": 10
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Login a user",
        "operationId": "user_auth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "8hf4fi3br8gd2828hny%&yhu38bub8c9c9uh744"
                    },
                    "user": {
                      "type": "object",
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email or password"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/logout": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Logs out a user",
        "operationId": "logout",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Access token revoked"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email or password"
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/subscriptions": {
      "get": {
        "tags": ["subscriptions"],
        "summary": "Get all the subscriptions in the DB",
        "operationId": "get_subscriptions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Create a subscription",
        "operationId": "create_subscription",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Subscription"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email or password"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/subscriptions/{subscription_id}": {
      "get": {
        "tags": ["subscriptions"],
        "summary": "Get a subscription by ID",
        "operationId": "get_subscription",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "description": "ID used to get subscription",
            "required": true,
            "schema": {
              "type": "string",
              "example": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["subscriptions"],
        "summary": "Update a subscription",
        "operationId": "update_subscription",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "description": "ID used to get subscription",
            "required": true,
            "schema": {
              "type": "string",
              "example": 100
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Subscription"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Failed to update subscription"
          }
        }
      },
      "delete": {
         "tags": [
          "subscriptions"
        ],
        "summary": "Deletes a subscription by ID",
        "operationId": "delete_subscription",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "description": "ID used to get subscription",
            "required": true,
            "schema": {
              "type": "string",
              "example": 10
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@gmail.com"
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "id": {
            "type": "string",
            "example": 10
          }
        },
        "xml": {
          "name": "user"
        }
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "10"
          },
          "subscription_name": {
            "type": "string",
            "example": "Hostinger"
          },
          "subscription_status": {
            "type": "boolean",
            "example": "true"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "The date the subscription started YYYY-MM-DD"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "description": "The date the subscription will end YYYY-MM-DD"
          },
          "users": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "james.bond@gmail.com"
            }
          },
          "subscription_cost": {
            "type": "integer",
            "example": 50000,
            "description": "The cost of the subscription"
          },
          "subscription_description": {
            "type": "string",
            "example": "Our web hosting provider subscription",
            "description": "A short text describing the subscription"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}