# server/api/v1/views/subscriptions.py
#!/usr/bin/python3
""" Subscription API endpoints """
from models.subscription import Subscription
from models.user import User
from models import storage
from api.v1.views import app_views
from flask_jwt_extended import jwt_required, get_jwt_identity
from flask import abort, jsonify, make_response, current_app, request, session
from flask_cors import cross_origin
from api.v1.email_service import send_welcome_email_task
from datetime import datetime

@app_views.route('/subscriptions', methods=['GET'], strict_slashes=False)
@jwt_required()
def get_subscriptions():
    print("WE ARE RUNNING")
    all_subscriptions = storage.all(Subscription).values()
    # print(all_subscriptions)
    list_subscriptions = []
    for subscription in all_subscriptions:
        list_subscriptions.append(subscription.make_api_response())
    return (jsonify(list_subscriptions)), 200

@app_views.route('/subscriptions/<subscription_id>', methods=['GET'],
        strict_slashes=False)
@jwt_required()
def get_subscription(subscription_id):
    subscription = storage.get(Subscription, subscription_id)
    if not subscription:
        abort(404)
    return jsonify(subscription.to_dict()), 200

@app_views.route('/subscriptions/my_subscriptions', methods=['GET'],
        strict_slashes=False)
@jwt_required()
def get_user_subscriptions():
    user_id = get_jwt_identity()
    try:
        subscriptions = storage.get_all_subscriptions_for_specific_user(user_id)
    except Exception as e:
        current_app.logger.critical("Could not get user subcriptions due to exception {e}")
        abort(404)
    return make_response(jsonify(subscriptions), 200)

@app_views.route('/subscriptions/<subscription_id>', methods=['DELETE'], strict_slashes=False)
@jwt_required()
def delete_subscription(subscription_id):
    subscription = storage.get(Subscription, subscription_id)
    if not subscription:
        abort(404)

    current_app.logger.critical(f"Subscription {subscription.to_dict()['subscription_name']} has been deleted")
    subscription.delete()
    storage.save()
    return make_response(jsonify({}), 204)

@app_views.route('/subscriptions/', methods=['POST'],
        strict_slashes=False)
@jwt_required()
def create_subscription():
    if not request.get_json():
        abort(400, description="Invalid JSON")

    if 'subscription_name' not in request.get_json():
        abort(400, description="Missing subscription name")
    if 'start_date' not in request.get_json():
        abort(400, description="Missing start date")
    if 'expiry_date' not in request.get_json():
        abort(400, description="Missing expiry date")
    if 'users' not in request.get_json():
        abort(400, description="Missing stakeholder emails")

    user_id = get_jwt_identity()
    request_data = request.get_json()
    subscription_creator = storage.get(User, user_id)
    print(subscription_creator)
    new_subscription = Subscription(subscription_creator, **request_data)
    new_subscription.save()
    send_welcome_email_task(new_subscription)
    subscription_response = new_subscription.make_api_response()
    current_app.logger.critical(f"Subscription {subscription_response['subscription_name']} has been created")
    print(subscription_response)
    return make_response(jsonify(subscription_response), 201)

@app_views.route('/subscriptions/<subscription_id>', methods=['PUT'],
        strict_slashes=False)
@jwt_required()
@cross_origin()
def update_subscription(subscription_id):
    subscription: Subscription = storage.get(Subscription, subscription_id)
    user_id = subscription.created_by
    if (user_id != get_jwt_identity()):
        current_app.logger.critical(f"User {user_id} attempted to illegally edit subscription {subscription_id}")
        abort(405, description="User not authorized to make request")
    if not subscription:
        abort(404)
    if not request.get_json():
        abort(400, description="Invalid JSON")
    data = request.get_json()
    try:
        print(data)
        subscription.update(data)
    except Exception as e:
        current_app.logger.critical(f"Updating subscription failed")
        current_app.logger.critical(f"Exception: {e}")
        abort(500, description="Failed to update subscription")
    current_app.logger.critical(f"Subscription {subscription.subscription_name} has been updated")
    subscription_response = subscription.make_api_response()
    response = make_response(jsonify(subscription_response), 200)
    return response