# server/models/subscription.py
#!/usr/bin/python3
"""Subscription model class declaration"""

from models.base_model import BaseModel , Base
import models
from sqlalchemy import Column, String, Boolean, DateTime, ForeignKey, Table, DECIMAL, Text
from datetime import datetime
from sqlalchemy.orm import relationship

user_subscriptions = Table('user_subscriptions', Base.metadata,
                          Column('user_id', String(60),
                                 ForeignKey('users.id', onupdate='CASCADE',
                                           ondelete='CASCADE'),
                                           primary_key=True),
                          Column('subscription_id', String(60),
                                 ForeignKey('subscriptions.id', onupdate='CASCADE',
                                            ondelete='CASCADE'),
                                            primary_key=True))

class Subscription(BaseModel, Base):
  "Subscription model class declaration"
  __tablename__ = 'subscriptions'
  subscription_name = Column(String(1024), nullable=False)
  subscription_status = Column(Boolean, nullable=False, default=True)
  subscription_cost = Column(DECIMAL, nullable=False, default=0.00)
  subscription_description = Column(Text, nullable=False)
  start_date = Column(DateTime, default=datetime.now(), nullable=False)
  expiry_date = Column(DateTime, nullable=False)
  last_notification = Column(DateTime, default=None)
  created_by = Column(String(60), ForeignKey('users.id', onupdate='CASCADE', ondelete='CASCADE'), nullable=False)
  users = relationship("User",
                       secondary=user_subscriptions,
                       viewonly=False)

  def __init__(self, user, *args, **kwargs):
    "Initializes Subscription"
    kwargs['created_by'] = user.id
    if kwargs['users']:
      stakeholders_array = []
      for user in kwargs['users']:
        stakeholder = models.storage.get_user_by_email(user)
        stakeholders_array.append(stakeholder)
      kwargs['users'] = stakeholders_array
      # print(kwargs)
    super().__init__(*args, **kwargs)
  
  def make_subscription_response(self):
    """ Formats a subscription object to only return necessay info to dashboard """
    subscription_dict = self.to_dict()
    users = models.storage.get_users_associated_with_a_subscription(self.id)
    # print("Users", users)
    result = {}
    necessary_keys = ['id',
                      'subscription_name',
                      'subscription_status',
                      'start_date',
                      'expiry_date',
                      'users',
                      'subscription_cost',
                      'subscription_description']
    for key in subscription_dict.keys():
      if key in necessary_keys:
        result[key] = subscription_dict[key]
    result['users'] = users
    return result